fix: Update the title on file changes
